requirements.txt
# BlackBook URL Scraping AI Agent Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# AI/ML
openai==1.3.0

# Data Processing
pydantic==2.5.0
python-dotenv==1.0.0

# Utilities
python-multipart==0.0.6
